Sort aggregate codegen
